+++ src/n1_lesson/Day.java
@@ -1,4 +1,4 @@
-package lessonOne;
+package n1_lesson;
 
 public enum Day {
     MONDAY,
