+++ src/n1_lesson/eq/readme.md
@@ -1,0 +1,38 @@
+Title: Object-Oriented Programming in Java - Advanced Concepts
+
+## Lesson Two: Advanced Java Programming
+
+### 1. Introduction to Object-Oriented Programming (OOP)
+- Recap of basic OOP principles: Encapsulation, Inheritance, and Polymorphism
+- Overview of the `UsefulObject` class as an example of encapsulation
+
+### 2. Class Design and Structure
+- Explanation of the `UsefulObject` class structure
+- Usage of private fields and public methods for encapsulation
+- The importance of providing meaningful getters and setters
+
+### 3. Equals and HashCode Methods
+- Understanding the `equals` method and its role in object comparison
+- Implementing the `hashCode` method for efficient hashing
+- Explanation of the `Objects` class for null-safe comparison
+
+### 4. Best Practices in Class Design
+- Overview of defensive programming practices
+- Proper usage of access modifiers (e.g., private, public)
+- Guidelines for implementing `equals` and `hashCode` methods
+
+### 5. Common Mistakes and Pitfalls
+- Identifying and avoiding common mistakes in class design
+- Handling null values and ensuring null safety
+- Best practices for designing classes that can be easily extended or modified
+
+### 6. Object Comparison and Immutability
+- Introduction to object comparison and the importance of overriding `equals` and `hashCode` for custom classes
+- Understanding immutability and its benefits in Java programming
+
+### 7. Advanced Java Concepts
+- Concepts of serialization and deserialization
+- Introduction to design patterns and their role in software architecture
+
+## Summary
+This lesson delves into advanced Java programming concepts with a focus on class design, encapsulation, and object comparison. The provided code for the `UsefulObject` class serves as a practical example, showcasing the implementation of `equals` and `hashCode` methods for proper object comparison. Best practices, common mistakes, and additional advanced topics are covered to enhance your understanding of Java's object-oriented features.
