+++ src/n1_lesson/readme.md
@@ -1,0 +1,44 @@
+Title: Java Basics and Advanced Concepts
+
+## Lesson One: Introduction to Java Programming
+
+### 1. Variables and Data Types
+- Overview of primitive data types: boolean, byte, short, char, int, float, long, double
+- Introduction to the 'var' keyword for type inference
+- Comparing primitives and understanding class wrappers (e.g., Integer)
+- String comparison and usage of the Comparator interface
+
+### 2. Control Flow and Switch Statements
+- Conditional statements: if-else, switch-case
+- Demonstration of a switch statement with both traditional and enhanced (JEP 361) syntax
+- Enumerations and the use of the 'Day' enum in switch statements
+
+### 3. Exception Handling
+- Introduction to try-catch blocks
+- Best practices for handling exceptions
+- Demonstration of try-with-resources for efficient resource management
+- Common mistakes in exception handling
+
+### 4. Collections in Java
+- Working with collections: List, Set, Map
+- Comparison between old and new ways of creating collections
+- Introduction to intermediate and terminal operations on streams
+- Overview of common stream operations and their usage
+
+### 5. Optional Class
+- Understanding the Optional class for handling potentially null values
+- Methods like of(), ofNullable(), isPresent(), orElse(), and orElseThrow()
+- Handling scenarios where null values may be encountered
+
+### 6. ExecutorService and Multithreading
+- Overview of ExecutorService and its role in managing thread pools
+- Demonstrating the usage of ExecutorService with a practical example
+- Best practices for managing concurrent tasks
+
+### 7. Additional Topics
+- Brief overview of file handling using Scanner and File classes
+- Introduction to the 'var' keyword for type inference
+- Basic concepts of lambda expressions and functional programming
+
+## Summary
+This lesson covered fundamental Java concepts such as variables, data types, control flow, exception handling, collections, Optional class, and multithreading using ExecutorService. The provided code includes practical examples to illustrate each concept, promoting a hands-on learning approach for Java programming.
