+++ src/n1_lesson/recordExample/readme.md
@@ -1,0 +1,65 @@
+### 1. Overview of Java Records
+Java Records, introduced in Java 16, are a concise way to define immutable data structures. They automatically generate methods such as `equals`, `hashCode`, and `toString`, reducing boilerplate code.
+
+### 2. Features of Java Records
+Records come with several features:
+- **Automatic Method Generation**: Records automatically generate `equals`, `hashCode`, and `toString` methods based on the record components.
+- **Immutability**: Record components are implicitly `final`, making records immutable by default.
+- **Conciseness**: Records provide a more concise syntax compared to traditional classes.
+
+### 3. Declaring Record Components
+Record components are the data fields within a record. In the `UserPost` record, we have three components: `user`, `postRid`, and `size`.
+
+```java
+public record UserPost(
+        String user,
+        String postRid,
+        int size
+) {}
+```
+
+### 4. Benefits of Java Records
+Let's compare a traditional Java class with the equivalent Java Record:
+
+Traditional Java Class:
+```java
+public class UserPost {
+    private String user;
+    private String postRid;
+    private int size;
+
+    public UserPost(String user, String postRid, int size) {
+        this.user = user;
+        this.postRid = postRid;
+        this.size = size;
+    }
+
+    // Getters, equals, hashCode, and toString methods manually implemented
+}
+```
+
+Equivalent Java Record:
+```java
+public record UserPost(String user, String postRid, int size) {}
+```
+
+The record version is significantly more concise, as it automatically generates the required methods.
+
+### 5. Use Cases and Limitations
+Java Records are beneficial for representing simple data structures. However, they might not be suitable for more complex scenarios or when mutability is required. Use records when simplicity and immutability are key.
+
+### 6. Migrating from Classes to Records
+To migrate an existing class to a record, simply replace the class definition with the record definition. Ensure that the class follows the requirements for records, such as immutability.
+
+### 7. Advanced Record Features
+Java Records support features like sealing and non-sealing. Sealing restricts which classes can be subclasses, enhancing control over the inheritance hierarchy. Non-sealing allows any class to extend the record.
+
+```java
+public sealed record SealedUserPost(String user, String postRid, int size)
+        permits AdminPost, RegularPost {}
+
+public final record AdminPost(String user, String postRid, int size) extends SealedUserPost {}
+public final record RegularPost(String user, String postRid, int size) extends SealedUserPost {}
+```
+
+In this example, `SealedUserPost` is a sealed record that only allows subclasses `AdminPost` and `RegularPost`.
